refactor: change file name of useThemeStore
