+++ src/App.tsx
@@ -1,9 +1,8 @@
 import { ThemeProvider } from '@emotion/react';
 import { GlobalStyle, theme } from 'styles';
 
-import { useThemeStore } from '@/stores';
-
 import Home from '@/components/Home';
+import { useThemeStore } from '@/stores/useThemeStore';
 
 function App() {
   const mode = useThemeStore((state) => state.theme);
